requirements.txt
google-adk==1.2.1
google-generativeai
google-cloud-bigquery